junklib: fix a regression causing id3v2 writing to crash when junk_iconv returns -1 (fixed #2396)

+++ junklib.c
@@ -2041,9 +2041,9 @@ junk_id3v2_add_text_frame2 (DB_id3v2_tag_t *tag, const char *frame_id, const cha
 
     int encoding = 0;
 
-    size_t outlen = -1;
+    int outlen = -1;
     if (tag->version[0] == 4) {
-        outlen = inlen;
+        outlen = (int)inlen;
         out = (uint8_t *)value;
         encoding = 3;
     }
@@ -2131,13 +2131,13 @@ junk_id3v2_add_comment_frame (DB_id3v2_tag_t *tag, const char *frame_id, const c
     char *buffer = malloc (buffersize);
 
     int enc = 0;
-    size_t l;
+    int l;
 
     if (tag->version[0] == 4) {
         // utf8
         enc = 3;
         memcpy (buffer, input, inputsize);
-        l = inputsize;
+        l = (int)inputsize;
     }
     else {
         l = junk_iconv (input, (int)inputsize, buffer, (int)buffersize, UTF8_STR, "cp1252");
@@ -2239,10 +2239,10 @@ junk_id3v2_add_txxx_frame (DB_id3v2_tag_t *tag, const char *key, const char *val
     int encoding = 0;
 
 
-    size_t res;
+    int res;
 
     if (tag->version[0] == 4) {
-        res = len;
+        res = (int)len;
         encoding = 3;
         memcpy (out, key, keylen);
         out[keylen] = 0;
